XMR<>BTC integration
Integration for sending Bitcoin to receive Monero.
The current implementation is completely detached from the actual Automated Swap Backend (ASB) that acts as the swap provider.
Since the rate of the ASB is effectively Kraken rate from Kraken, the current XMR-BTC rate uses the Kraken ticker websocket to determine the current price.
This solution is not optimal, but we ran into trouble with fetching the rate over Tor because currently opendex is not hosted via an onion address. We will tackle the rate fetching from the actual ASB again once that is sorted out.