Allow single direction for swap-provider pair
At the moment COMIT can only support buying Monero.
This is due to the nature of the XMR<>BTC protocol - where Bitcoin has to lock first, which means the service provider should never be on the Bitcoin lock side.
The previous swap-provider and currency pair relation only allowed supporting the currency pair fully (i.e. in both directions) or not.
With this change we can now depict supporting only one direction for a currency pair.
I tried to keep it simple to extend adding pair support for other providers and default to support both directions.